LPS-54323 move classes only
