+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/portlet/PasswordPoliciesAdminPortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.passwordpoliciesadmin;
+package com.liferay.password.policies.admin.web.portlet;
 
 import com.liferay.portal.DuplicatePasswordPolicyException;
 import com.liferay.portal.NoSuchPasswordPolicyException;
