+++ modules/apps/password-policies-admin/password-policies-admin-web/src/com/liferay/password/policies/admin/web/search/UserPasswordPolicyChecker.java
@@ -12,25 +12,25 @@
  * details.
  */
 
-package com.liferay.portlet.passwordpoliciesadmin.search;
+package com.liferay.password.policies.admin.web.search;
 
 import com.liferay.portal.kernel.dao.search.RowChecker;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.model.Organization;
 import com.liferay.portal.model.PasswordPolicy;
 import com.liferay.portal.model.PasswordPolicyRel;
-import com.liferay.portal.service.OrganizationLocalServiceUtil;
+import com.liferay.portal.model.User;
 import com.liferay.portal.service.PasswordPolicyRelLocalServiceUtil;
+import com.liferay.portal.service.UserLocalServiceUtil;
 
 import javax.portlet.RenderResponse;
 
 /**
  * @author Scott Lee
  */
-public class OrganizationPasswordPolicyChecker extends RowChecker {
+public class UserPasswordPolicyChecker extends RowChecker {
 
-	public OrganizationPasswordPolicyChecker(
+	public UserPasswordPolicyChecker(
 		RenderResponse renderResponse, PasswordPolicy passwordPolicy) {
 
 		super(renderResponse);
@@ -40,12 +40,11 @@ public class OrganizationPasswordPolicyChecker extends RowChecker {
 
 	@Override
 	public boolean isChecked(Object obj) {
-		Organization organization = (Organization)obj;
+		User user = (User)obj;
 
 		try {
-			return OrganizationLocalServiceUtil.hasPasswordPolicyOrganization(
-				_passwordPolicy.getPasswordPolicyId(),
-				organization.getOrganizationId());
+			return UserLocalServiceUtil.hasPasswordPolicyUser(
+				_passwordPolicy.getPasswordPolicyId(), user.getUserId());
 		}
 		catch (Exception e) {
 			_log.error(e, e);
@@ -56,13 +55,12 @@ public class OrganizationPasswordPolicyChecker extends RowChecker {
 
 	@Override
 	public boolean isDisabled(Object obj) {
-		Organization organization = (Organization)obj;
+		User user = (User)obj;
 
 		try {
 			PasswordPolicyRel passwordPolicyRel =
 				PasswordPolicyRelLocalServiceUtil.fetchPasswordPolicyRel(
-					Organization.class.getName(),
-					organization.getOrganizationId());
+					User.class.getName(), user.getUserId());
 
 			if ((passwordPolicyRel != null) &&
 				(passwordPolicyRel.getPasswordPolicyId() !=
@@ -79,7 +77,7 @@ public class OrganizationPasswordPolicyChecker extends RowChecker {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		OrganizationPasswordPolicyChecker.class);
+		UserPasswordPolicyChecker.class);
 
 	private final PasswordPolicy _passwordPolicy;
